// frontend/src/models/manualPdf.model.ts
export class ManualPdfData {
    medicalRecordNo: string;
    SID: string;
    departmentNo: string;
    checkDate: string;
    patientBirth : string;
    specimenNo: string;
    patientName: string;
    patientSex : string;
    specimenType : string;
    specimenStatus : string;

    checker: number;
    qualityManager: number;
    reportDoctor: number;
    confirmer: number;

}